fix/chore: update documentation of the file format
Having a specification of the file format is great and really helpful!
The current version is sometimes imprecise, sometimes incomplete
(e.g. dangling sentences) and sometimes outdated. Correct all three.

In addition, remove (any implicit mention of) the 'identifiers' field from
the README: these are more unstable; we prefer to track stable URLs
(which point to the respective theorems) rather than identifiers which become
outdated more quickly.

Found while implementing a more typed parser for this, for use in synchronising
this database with a file of formalisations in mathlib.