chore: update Lean formalisations (#11)
This is automatically generated from a .yaml file in mathlib. A script
to generate this commit will be added in a future pull request.

I am not particularly fond of the new yaml file formatting. When the
script is PRed, I'll gladly take tweaks to the script which preserve the
old formatting.