Make rows hyperlinkable (#25)
Sets the HTML identifier of each row to its Wikidata identifier, and
accepts page length as a GET parameter. This is a bit of a hacky
solution, but it makes it possible to link to a specific row by using a
link of the form `https://1000-plus.github.io/?length=-1#Q3527263` or
`https://1000-plus.github.io/all.html?length=-1#Q3527263`.

Resolves #24.